py/saga/requirements.txt
beautifulsoup4==4.6.3
lxml==4.2.4
Flask==1.0.2
Flask-Script==2.0.6
Janome==0.3.7
requests==2.21.0
pytest==4.0.2
Pillow==5.3.0
flake8>=3.6.0
autopep8>=1.4.3
PyYAML>=3.13
